Add credentials structure validation logic
Add email and password regex matching so that the login input takes a standard
form. Display a snackbar when the user presses "done" to notify the user
if the email is not valid to increase app login clarity.

Update activity lifecycle logs so that they carry more information.

Update certain xml code to help with the mentioned functionality.