Implement taking profile pictures
Make it possible to only open one instance of the profile dialog at a
time.
Add a similar functionality to the dialog in show details. The only
issue is that the state of the last review added is remembered upon
opening the dialog, however in the future the user should be able to write
only one review so this will be ok.