Implement view model
Refactor code, a lot had to be adjusted for the code to be more readable
and work with observers at the same time.
Move show list and reviews map to respective viewModels.

Adjust the way showsFragment gets the data for email and password. It is
no longer taken from preferences and is now transferred using safe args.
The oversight was that nothing is saved to user preferences if the user
does not check remember me.